+++ src/main/java/model/User.java
@@ -5,6 +5,7 @@
 package model;
 
 import java.io.Serializable;
+import java.math.BigInteger;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
@@ -24,9 +25,9 @@ public class User implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Column(name = "id")
-    private int id;
+    private BigInteger id;
     @Column(name = "team_id")
-    private int teamId;
+    private BigInteger teamId;
     @Column(name = "firstName")
     private String firstName;
     @Column(name = "lastName")
@@ -34,19 +35,19 @@ public class User implements Serializable {
     @Column(name = "email")
     private String email;
             
-    public int getId() {
+    public BigInteger getId() {
         return id;
     }
     
-    public void setId(int id) {
+    public void setId(BigInteger id) {
         this.id = id;
     }
     
-    public int getTeamId() {
+    public BigInteger getTeamId() {
         return teamId;
     }
     
-    public void setTeamId(int teamId) {
+    public void setTeamId(BigInteger teamId) {
         this.teamId = teamId;
     }
     
@@ -77,7 +78,7 @@ public class User implements Serializable {
     @Override
     public int hashCode() {
         int hash = 0;
-        hash += (int) id;
+        hash += id.intValue();
         return hash;
     }
